refactor(core): group text helpers under core text
Move CJK joining, semantic spacing, and text-direction utilities into a text domain folder so the core root keeps moving toward stable entrypoints plus categorized internals.
