+++ src/core/text/cjkJoin.ts
