+++ src/core/text/spacing.ts
