+++ src/core/text/textDirection.ts
